reorganized api routes
