+++ src/pages/api/lib/math/prooflike/index.ts
@@ -1,16 +1,11 @@
 import type { NextApiRequest, NextApiResponse } from "next";
-import { findScienceImagesProblemById } from "../../../../../server/mongodb/actions/mcatProblem";
+import { getAllMathProofProblems } from "../../../../../server/mongodb/actions/numericalProblem";
 import requestWrapper from "../../../../../server/utils/middleware";
 
 async function handler(req: NextApiRequest, res: NextApiResponse) {
-  if (req.method !== "GET") {
-    res.status(405).end();
-    return;
-  }
-  const { id } = req.query;
-  const problem = await findScienceImagesProblemById(id);
+  const allProblems = await getAllMathProofProblems();
 
-  const formattedResponse = JSON.stringify(problem, null, 2);
+  const formattedResponse = JSON.stringify(allProblems, null, 2);
 
   res.setHeader("Content-Type", "application/json");
   res.status(200).send(formattedResponse);
