+++ src/pages/api/lib/math/symbolic/index.ts
@@ -1,0 +1,14 @@
+import type { NextApiRequest, NextApiResponse } from "next";
+import { getAllMathSymbolicProblems } from "../../../../../server/mongodb/actions/numericalProblem";
+import requestWrapper from "../../../../../server/utils/middleware";
+
+async function handler(req: NextApiRequest, res: NextApiResponse) {
+  const allProblems = await getAllMathSymbolicProblems();
+
+  const formattedResponse = JSON.stringify(allProblems, null, 2);
+
+  res.setHeader("Content-Type", "application/json");
+  res.status(200).send(formattedResponse);
+}
+
+export default requestWrapper(handler, "GET");
