+++ src/pages/api/lib/mcatScience/img/val/[id].ts
@@ -1,6 +1,6 @@
 import type { NextApiRequest, NextApiResponse } from "next";
-import { findPhysicsProblemById } from "../../../../../server/mongodb/actions/numericalProblem";
-import requestWrapper from "../../../../../server/utils/middleware";
+import { findScienceImagesProblemById } from "../../../../../../server/mongodb/actions/mcatProblem";
+import requestWrapper from "../../../../../../server/utils/middleware";
 
 async function handler(req: NextApiRequest, res: NextApiResponse) {
   if (req.method !== "GET") {
@@ -8,7 +8,7 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
     return;
   }
   const { id } = req.query;
-  const problem = await findPhysicsProblemById(id);
+  const problem = await findScienceImagesProblemById(id);
 
   const formattedResponse = JSON.stringify(problem, null, 2);
