+++ src/pages/api/lib/mcatScience/test/[id].ts
@@ -1,11 +1,16 @@
 import type { NextApiRequest, NextApiResponse } from "next";
-import { getAllPhysicsProblems } from "../../../../../server/mongodb/actions/numericalProblem";
+import { findScienceTestProblemById } from "../../../../../server/mongodb/actions/mcatProblem";
 import requestWrapper from "../../../../../server/utils/middleware";
 
 async function handler(req: NextApiRequest, res: NextApiResponse) {
-  const allProblems = await getAllPhysicsProblems();
+  if (req.method !== "GET") {
+    res.status(405).end();
+    return;
+  }
+  const { id } = req.query;
+  const problem = await findScienceTestProblemById(id);
 
-  const formattedResponse = JSON.stringify(allProblems, null, 2);
+  const formattedResponse = JSON.stringify(problem, null, 2);
 
   res.setHeader("Content-Type", "application/json");
   res.status(200).send(formattedResponse);
