+++ src/pages/api/lib/physics/symbolic/img/[id].ts
@@ -1,6 +1,6 @@
 import type { NextApiRequest, NextApiResponse } from "next";
-import { findPhysicsImgProblemById } from "../../../../../server/mongodb/actions/numericalProblem";
-import requestWrapper from "../../../../../server/utils/middleware";
+import { findPhysicsImgTestProblemById } from "../../../../../../server/mongodb/actions/numericalProblem";
+import requestWrapper from "../../../../../../server/utils/middleware";
 
 async function handler(req: NextApiRequest, res: NextApiResponse) {
   if (req.method !== "GET") {
@@ -8,7 +8,7 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
     return;
   }
   const { id } = req.query;
-  const problem = await findPhysicsImgProblemById(id);
+  const problem = await findPhysicsImgTestProblemById(id);
 
   const formattedResponse = JSON.stringify(problem, null, 2);
